Unified fit Slit smeared data and Smearing improvements
Fixed Unified fit so it does not require Qmax for fitting to be 3xSlit
length. Same fix as in Modeling II - if needed, Qmax for modeling is
extended and after modeling and smearing truncated again.
Increased speed of smearing data about 6x, which improves speed of all
tools with smeared data as well as Desmearing.